fixing jobs rendering with expire date
problem: not all jobs are showing up
solution: the bug is inconsitent rendering of the expires date,
which depending on how it is provided can be rendered correctly
as a timestamp, or if jekyll cannot, it falls back to a date format
that cannot pass the comparison. In practice this results in a lot
of missing jobs, and introduces potential for many errors. A better
solution is to rely on what we already have, parsing the dates and
expires in the python automation, and then allowing for 60 days
of buffer (so jobs are shown that go 60 days back). To allow for this
we have updated the "Expires" heading to be "Advertised until"

Signed-off-by: vsoch <vsoch@users.noreply.github.com>